Allow more-recent versions of Rest Client gem
Release notes:
https://github.com/rest-client/rest-client/blob/master/history.md

Elastcitiy uses `RestClient` only in `Elasticity:AwsSession#submit`, and
none of the changes in the release notes seem relevant to these uses.

James P forked this earlier and made the same changed, but it was based
on a more-recent version of the gem that was incompatible with the
version of Fog we use in academia-app (and I don't want to have to
upgrade that as well).